controllers done v1

+++ app/Http/Controllers/Sub_categoriesController.php
@@ -2,27 +2,20 @@
 
 namespace App\Http\Controllers;
 use App\Sub_category;
+use App\Product;
 
 use Illuminate\Http\Request;
 
 class Sub_CategoriesController extends Controller
 {
-
-	public function index($categoryId)
-	{
-        	if (!$subCategories = Sub_category::where('categoryId', $categoryId)->get()) {
-				abort(404);
-			}
-
-			return view('partials.subcategories', array('subCategories'=>$subCategories));
-    }
-
+	//for creating a new subcategory with a form
 	public function create()
 	{
 		//llamar a vista de agregar subcategory
 
 	}
 
+	//for editing a subcategory loading its data in the form
 	public function edit($subCategoryId)
 	{
 		if (!$subCategory = Sub_category::find($subCategoryId)) {
@@ -33,7 +26,8 @@ class Sub_CategoriesController extends Controller
 
 	}
 
-	public function store(Sub_category $request) { //recibira un array asociativo con los datos de la subcategoria
+	//it receives the data of a new subcategory from a form and save it in the database
+	public function store(Sub_category $request) {
 		$subCategory = new Sub_category();
 
 		$subCategory->name = $request->input('name');
@@ -43,6 +37,7 @@ class Sub_CategoriesController extends Controller
 		$subCategory->save();
 	}
 
+	//it looks for products where subcategoryid is this and call the view
 	public function show($subCategoryId)
 	{
         if (!$products = Product::where('subCategoryId', $subCategoryId)->get()) {
@@ -52,7 +47,8 @@ class Sub_CategoriesController extends Controller
 		return view('partials.products', array('products'=>$products));
     }
 
-	public function update(Sub_category $request, $subCategoryId, $auxSubCategory)
+	public function update(Sub_category $request, $subCategoryId)
+	//it updates a subcategory in the database
 	{
 		if (!$subCategory = Sub_category::find($subCategoryId)) {
 			abort(404);
@@ -65,13 +61,19 @@ class Sub_CategoriesController extends Controller
 		$subCategory->save();
     }
 
+	//for deleting a subcategory
 	public function destroy($subCategoryId)
 	{
         if (!$subCategory = Sub_category::find($subCategoryId)) {
 			abort(404);
 		}
 
-		$subCategory->delete();
+		if (!$products = Product::where('subCategoryId', $subCategoryId)->get()) {
+			abort(404);
+		}
 
+		if ($products->delete()) {
+			$subCategory->delete();
+		}
     }
 }
